Move back to int

+++ valen-game-eternal/src/main/java/be/twofold/valen/game/eternal/reader/havokshape/HkOption.java
@@ -25,7 +25,7 @@ enum HkOption implements FlagEnum {
     }
 
     @Override
-    public long value() {
+    public int value() {
         return value;
     }
 }
